mv wrapper module to model_wrapper.
